Add parameter to run evaluation without brute_force

+++ first_problem/evaluator.py
@@ -14,39 +14,44 @@ def load_json(json_name):
 
     return data
 
-def evaluate(data, id: int, return_dict):
+def evaluate(data, id: int, return_dict, use_brute_force=True):
     print(f'started proccess {id}')
     n = data['n']
     players_count = data['players_count']
     specters_count = data['specters_count']
     players_points = data['players_points']
     specters_points = data['specters_points']
-    print('hey', np.array(players_points).shape, players_count)
-    naive = brute_force_sol(n,players_count,specters_count,players_points,specters_points)
-    print(f'finished brute force {id}')
     hungarian = solve(n,players_count,specters_count,players_points,specters_points)
+    print(f'finished hungarian {id}')
+
+    if use_brute_force:
+        naive = brute_force_sol(n,players_count,specters_count,players_points,specters_points)
+        if hungarian != naive:
+            print()
+            print(f'hungarian {hungarian}, naive {naive}')
+            print(n)
+            print(players_count)
+            print(specters_count)
+            print(players_points)
+            print(specters_points)
+            print()
+        return_dict[id] = (naive, hungarian, hungarian==naive)
+    # print(f'finished brute force {id}')
+    
+    return_dict[id] = hungarian
     print(f'finished proccess {id}')
-    if hungarian != naive:
-        print()
-        print(f'hungarian {hungarian}, naive {naive}')
-        print(n)
-        print(players_count)
-        print(specters_count)
-        print(players_points)
-        print(specters_points)
-        print()
-    return_dict[id] = (naive, hungarian, hungarian==naive)
+    
     
 
 
-def run_evaluator(json_name, dump_to_json=True):
+def run_evaluator(json_name, dump_to_json=True, use_brute_force=True):
     data = load_json(json_name)
     print('started')
     manager = multiprocessing.Manager()
     return_dict = manager.dict()
     jobs = []
     for i in range(len(data)):
-        p = multiprocessing.Process(target=evaluate, args=(data[i], i, return_dict))
+        p = multiprocessing.Process(target=evaluate, args=(data[i], i, return_dict, use_brute_force))
         jobs.append(p)
         p.start()
 
@@ -54,24 +59,27 @@ def run_evaluator(json_name, dump_to_json=True):
         proc.join()
     
     values_list = list(return_dict.values())
+    if use_brute_force:
+        final_list = []
+        passed = 0
+        
+        for i in range(len(return_dict.values())):
+            if int(values_list[i][2]) == 1:
+                passed+=1
+            final_list.append({
+                'naive': int(values_list[i][0]),
+                'hungarian': int(values_list[i][1]),
+                'Equal': int(values_list[i][2])
 
-    final_list = []
-    passed = 0
-    for i in range(len(return_dict.values())):
-        if int(values_list[i][2]) == 1:
-            passed+=1
-        final_list.append({
-            'naive': int(values_list[i][0]),
-            'hungarian': int(values_list[i][1]),
-            'Equal': int(values_list[i][2])
-
-        })
+            })
+        
+        print(f'passed {passed} cases of {len(data)}')
+        if dump_to_json:
+            json.dump(final_list, open('results.json', 'w'))
+        return final_list
     
-    print(f'passed {passed} cases of {len(data)}')
-    if dump_to_json:
-        json.dump(final_list, open('results.json', 'w'))
-    return final_list
+    print(values_list)
 
 
 
-run_evaluator('test_cases.json')
+run_evaluator('test_cases.json', use_brute_force=False)
